Move test files into __test__ directory
